+++ data/plugin-meta.php
@@ -5,7 +5,7 @@ return <<<'JSON'
     "namespace": "MWP\\Framework",
     "slug": "mwp-framework",
     "name": "MWP Application Framework for WordPress",
-    "description": "Provides an object oriented utility framework for building plugins/applications using WordPress.",
+    "description": "Provides an object oriented utility framework for building plugins\/applications using WordPress.",
     "author": "Kevin Carwile",
     "author_url": "http:\/\/www.millermedia.io\/",
     "version": "2.0.0",